refactor: completely deprecate field model sync
